+++ ibis/backends/dask/tests/execution/test_arrays.py
@@ -55,9 +55,7 @@ def test_array_collect(t, df):
     tm.assert_frame_equal(result.compute(), expected.compute())
 
 
-@pytest.mark.xfail(
-    raises=NotImplementedError, reason='TODO - windowing - #2553'
-)
+@pytest.mark.backends_notimpl(["dask"], reason="windowing - #2553")
 def test_array_collect_rolling_partitioned(t, df):
     window = ibis.trailing_window(1, order_by=t.plain_int64)
     colexpr = t.plain_float64.collect().over(window)
@@ -77,7 +75,7 @@ def test_array_collect_rolling_partitioned(t, df):
 
 
 # Need an ops.ArraySlice execution func that dispatches on dd.Series
-@pytest.mark.xfail(raises=NotImplementedError, reason="TODO - arrays - #2553")
+@pytest.mark.backends_notimpl(["dask"], reason="arrays - #2553")
 @pytest.mark.parametrize(
     ['start', 'stop'],
     [
@@ -152,42 +150,21 @@ def test_array_index_scalar(client, index):
     assert result == expected
 
 
-@pytest.mark.xfail(
-    raises=NotImplementedError,
-    reason='TODO - arrays - #2553'
-    # Need an ops.ArrayRepeat execution func that dispatches on dd.Series
-)
+@pytest.mark.backends_notimpl(["dask"], reason="arrays - #2553")
 @pytest.mark.parametrize('n', [1, 3, 4, 7, -2])  # negative returns empty list
 @pytest.mark.parametrize('mul', [lambda x, n: x * n, lambda x, n: n * x])
 def test_array_repeat(t, df, n, mul):
     expr = t.projection([mul(t.array_of_strings, n).name('repeated')])
-    result = expr.compile()
-    expected = dd.from_pandas(
-        pd.DataFrame({'repeated': df.array_of_strings * n}),
-        npartitions=1,
-    )
-    tm.assert_frame_equal(result.compute(), expected.compute())
+    result = expr.execute()
+    expected = pd.DataFrame({'repeated': df.array_of_strings * n})
+    tm.assert_frame_equal(result, expected)
 
 
-@pytest.mark.parametrize('n', [1, 3, 4, 7, -2])  # negative returns empty list
-@pytest.mark.parametrize('mul', [lambda x, n: x * n, lambda x, n: n * x])
-def test_array_repeat_scalar(client, n, mul):
-    raw_array = [1, 2]
-    array = ibis.literal(raw_array)
-    expr = mul(array, n)
-    result = client.execute(expr)
-    expected = mul(raw_array, n)
-    assert np.array_equal(result, expected)
-
-
-@pytest.mark.xfail(
-    raises=ValueError,
-    reason='TODO - arrays - #2553'
-    # ValueError: Dask backend borrows Pandas backend's Cast execution
-    # function, which assumes array representation is np.array.
-    # NotImplementedError: Need an ops.ArrayConcat execution func that
-    # dispatches on dd.Series
-)
+# ValueError: Dask backend borrows Pandas backend's Cast execution
+# function, which assumes array representation is np.array.
+# NotImplementedError: Need an ops.ArrayConcat execution func that
+# dispatches on dd.Series
+@pytest.mark.backends_notimpl(["dask"], reason="arrays - #2553")
 @pytest.mark.parametrize('op', [lambda x, y: x + y, lambda x, y: y + x])
 def test_array_concat(t, df, op):
     x = t.array_of_float64.cast('array<string>')
