+++ ibis/backends/dask/tests/execution/test_cast.py
@@ -77,7 +77,7 @@ def test_cast_timestamp_column(t, df, column, to, expected):
         param(
             'double',
             float,
-            marks=pytest.mark.xfail(raises=NotImplementedError),
+            marks=pytest.mark.backends_notimpl(["dask"]),
         ),
         (
             dt.Timestamp('America/Los_Angeles'),
@@ -98,11 +98,7 @@ def test_cast_timestamp_scalar_naive(to, expected):
     [
         ('string', str),
         ('int64', lambda x: x.value),
-        param(
-            'double',
-            float,
-            marks=pytest.mark.xfail(raises=NotImplementedError),
-        ),
+        param('double', float, marks=pytest.mark.backends_notimpl(["dask"])),
         (
             dt.Timestamp('America/Los_Angeles'),
             lambda x: x.tz_convert('America/Los_Angeles'),
