+++ ibis/backends/dask/tests/execution/test_strings.py
@@ -93,10 +93,8 @@ from pytest import param
             lambda s: s.split(' '),
             lambda s: s.str.split(' '),
             id='split_spaces',
-            marks=pytest.mark.xfail(
-                raises=NotImplementedError,
-                reason='TODO - arrays - #2553'
-                # Need ops.StringSplit exec func that dispatches on dd.Series
+            marks=pytest.mark.backends_notimpl(
+                ["dask"], reason='arrays - #2553'
             ),
         ),
     ],
