+++ ibis/backends/tests/test_api.py
@@ -23,7 +23,7 @@ def test_exists_table(con, table_name):
 # 1. `current_database` returns '.', but isn't listed in list_databases()
 # 2. list_databases() returns directories which don't make sense as HDF5
 #    databases
-@pytest.mark.backends_never(["dask", "pandas"])
+@pytest.mark.backends_never(["dask", "pandas"], reason="pass")
 @pytest.mark.backends_notimpl(["datafusion"])
 def test_database_consistency(con):
     # every backend has a different set of databases, not testing the
